Revert "redesign architecture"
