--- Application - Copy/.idea/libraries/sqlite1.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="sqlite1">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/spellchecker/spellchecker/sqlite/sqlite.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
