--- Application - Copy/spellchecker/spellchecker/interfaces/ICheckable.java
@@ -1,7 +1,0 @@
-package spellchecker.interfaces;
-
-import spellchecker.models.Database;
-
-public interface ICheckable {
-    boolean check(IDatabase database);
-}
