--- Application - Copy/spellchecker/spellchecker/interfaces/IDatabase.java
@@ -1,5 +1,0 @@
-package spellchecker.interfaces;
-
-public interface IDatabase {
-    boolean checkWord(String text);
-}
