--- Application - Copy/spellchecker/spellchecker/interfaces/IViewInterface.java
@@ -1,7 +1,0 @@
-package spellchecker.interfaces;
-
-public interface IViewInterface {
-    void start();
-    String getTextInput();
-    void updateResultLabel(boolean value);
-}
